# src/server.js
import { UPLOAD_DIR } from './constants/index.js';
import cookieParser from 'cookie-parser';
import express from 'express';
import pino from 'pino-http';
import cors from 'cors';
/* import contactsRouter from './routers/contacts.js'; */
import router from './routers/index.js';

import { getEnvVar } from './utils/getEnvVar.js';
import { errorHandler } from './middlewares/errorHandler.js';
import { notFoundHandler } from './middlewares/notFoundHandler.js';
/* import { getAllContacts, getContactById } from './services/contacts.js'; */
import { swaggerDocs } from './middlewares/swaggerDocs.js';
export const setupServer = () => {
  const PORT = Number(getEnvVar('PORT', '3000'));
  const app = express();
  app.use(express.json());
  app.use(cors());
  app.use(cookieParser());

  app.use(express.json());
  app.use(cors());

  app.use(
    pino({
      transport: {
        target: 'pino-pretty',
      },
    }),
  );

  app.get('/', (req, res) => {
    res.json({
      message: 'Hello MondoDB!',
    });
  });

  /* app.use(contactsRouter); */
  app.use(router);
  app.use('/uploads', express.static(UPLOAD_DIR));
  app.use('/api-docs', swaggerDocs());
  app.use(notFoundHandler);
  app.use(errorHandler);
  /* app.get('/contacts', async (req, res) => {
    const contacts = await getAllContacts();
    res.status(200).json({
      message: 'Successfully found contacts!',
      data: contacts,
    });
  });

  app.get('/contacts/:contactId', async (req, res, next) => {
    const { contactId } = req.params;
    const contact = await getContactById(contactId);
    if (!contact) {
      res.status(404).json({
        message: 'Contact not found',
      });
      return;
    }
    res.status(200).json({
      message: `Successfully found contact with id ${contactId}!`,
      data: contact,
    });
  }); */

  /* app.use((req, res, next) => {
    res.status(404).json({
      message: 'Not found',
    });
  }); */
  /* app.use((err, req, res, next) => {
    res.status(500).json({
      message: 'Something went wrong',
      error: err.message,
    });
  }); */

  app.listen(PORT, () => {
    console.log(`Server is running on port ${PORT}`);
  });
};